backend/transcription-service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
motor==3.3.1
pymongo==4.5.0
pydantic==1.10.13
numpy==1.24.3
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.0.1+cpu
openai-whisper==20230918 